requirements.txt
torch==2.0.0
tqdm==4.64.1
numpy==1.23.5
matplotlib==3.7.1
scikit-learn==1.2.2
PyYAML==6.0
torch-geometric==2.2.0
transformers==4.30.2
wandb==0.15.4